copy function update, complete 80%

+++ src/com/cha122977/android/filecontroller/FileControllerActivity.java
@@ -43,13 +43,9 @@ public class FileControllerActivity extends Activity {
         initial();
         setListener();
         
-//        File f = Environment.getExternalStorageDirectory
-        File f = new File("/sdcard/");
-        String x = f.getAbsolutePath();
-        String y = f.getPath();
+        openTopFile(ROOT);//at initial: open the SD-card root
+        openBottomFile(ROOT);//at initial: open the SD-card root
         
-        openTopFile(x);//at initial: open the SD-card root
-        openBottomFile(y);//at initial: open the SD-card root
     }
     private void setView(){
     	tv_topDir = (TextView)findViewById(R.id.topTextView);
@@ -219,38 +215,46 @@ public class FileControllerActivity extends Activity {
     	//TODO open a dialog and a editText to write new name.
     	
     }
-    private void copyFile(String copieer, String target){ 
+    private void copyFile(String copieer, String target){ //copy file to target(directory) as same name.
     	InputStream in;
     	OutputStream out;
     	//find the file name of copied file name
     	int nameIndex = copieer.lastIndexOf("/");
     	String copieerFileName = copieer.substring(nameIndex+1);
     	//TODO Avoid replace the presence data which have the same file name in target directory
-    	try{
-    		Log.d("TAG", "Copieer: " + copieer +", Target: " + target);
-    		in = new FileInputStream(copieer);
-    		Log.d("TAG", "Created in Succeed");
-    		out = new FileOutputStream(target + "/" + copieerFileName);
-    		Log.d("TAG", "Created out Succeed");
+    	String completeTargetFilePath = target + "/" + copieerFileName;
+    	if(new File(completeTargetFilePath).exists() == false){//there is no file have same name at target path.
+	    	try{
+	    		Log.d("TAG", "Copieer: " + copieer +", Target: " + target);
+	    		in = new FileInputStream(copieer);
+	    		Log.d("TAG", "Created in Succeed");
+	    		out = new FileOutputStream(target + "/" + copieerFileName);
+	    		Log.d("TAG", "Created out Succeed");
+	    		
+		    	byte[] buffer = new byte[1024];
+		        int read;
+		        while((read = in.read(buffer)) != -1){
+		          out.write(buffer, 0, read);
+		        }
+		        Log.d("TAG", "write data over");
+		        in.close();
+		        out.flush();
+		        out.close();
+		        Log.d("TAG", "Copy File Succeed");
+		        
+		        openBottomFile(target);//use to refresh data
+	    	} catch(IOException e){
+	    		Log.d("TAG", "Copy File Error");
+	    		Toast.makeText(getApplicationContext(), "Copy File Error", Toast.LENGTH_LONG);
+	    	} finally{
+	    		in = null;
+	    		out = null;
+	    	}
+    	} else {//have same name file.
+    		//TODO open an alert have item: 1. still copy, replace the file; 
+    		//								2. still copy, but change name; (new name: [old name + "(x)"] ), x is {1,2,3....} 
+    		//								3. cancel.
     		
-	    	byte[] buffer = new byte[1024];
-	        int read;
-	        while((read = in.read(buffer)) != -1){
-	          out.write(buffer, 0, read);
-	        }
-	        Log.d("TAG", "write data over");
-	        in.close();
-	        out.flush();
-	        out.close();
-	        Log.d("TAG", "Copy File Succeed");
-	        
-	        openBottomFile(target);//use to refresh data
-    	} catch(IOException e){
-    		Log.d("TAG", "Copy File Error");
-    		Toast.makeText(getApplicationContext(), "Copy File Error", Toast.LENGTH_LONG);
-    	} finally{
-    		in = null;
-    		out = null;
     	}
     }
     //------------File Option function/>-----//
